SemanticRole/requirements.txt
django==3.2
djangorestframework
markdown
django-filter
django-cors-headers
psycopg2-binary
